requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# LLM Backends (install what you need)
# For Ollama (recommended for local/free)
ollama>=0.1.0

# For Anthropic Claude (paid)
anthropic>=0.25.0

# For Groq (free tier available)
groq>=0.4.0

# For HuggingFace (free tier available)
huggingface-hub>=0.19.0

# Image Generation Backends (install what you need)
# For DALL-E (OpenAI)
openai>=1.0.0

# For Replicate (free tier)
replicate>=0.15.0

# For Segmind and general HTTP
requests>=2.31.0

# Optional: for better async support
aiofiles>=23.0.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0